Fixed the magic-number delay
As you warned in comments, it was constantly a low level in this place
for my hardware. For me the delay is 70, not 50. So I advice you to
avoid magic numbers and check the level directly in an endless cycle.
Please decorate it with some error code if you like, but the original
gode does not work. It took an evening for me to figure up why.