Add an arg parser to simpletest.py
This change adds an argument parser so that you can call simpletest.py
directly from the command line.  You can invoke it with:

$ python simpletest.py --sensor <sensor type> --pin <pin #>

Sensor can be one of 'dht11', 'dht22', or 'am2302', however defaults to
the dht22 sensor.  Pin can either be a string or an integer and
corresponds to the GPIO pin that is attached to the sensor.