Fix egg_info and sdist build issue on MS Windows
The platform detect code is run even if a binary is not being built,
however it relies on /proc/cpuinfo being present. If it is not, platform_detect
will fail with a FileNotFoundError rather than returning UNKNOWN. This
change makes it possible to run the non-binary commands on operating systems
other than Linux. The motivation for this is that Pipenv requires egg_info
to be executable to determine dependencies, even if the package is question
is specified as a conditional dependency whose conditions are not met on
the current host.